Revert to older TypeORM version
I had to upgrade TypeORM to an alpha version because it had the necessary fixes to run the migrations and populate the database color palettes table. Unfortunately, it looks like that version was causing increased CPU load and brought the API down. I deployed an older version of the API manually.