Add notablePeopleEvents query, use uppercase for most enum types (#102)
This PR includes some (technically) breaking changes to the enum types. The enum values for enum types that are not yet used on the production website are now all uppercase for consistency.

There remains a few non-uppercase enum types that cannot be changed now because they would break the website and they are stored in lowercase in the database so we'll need to figure out how to make those consistent with the other enum types.

After we complete the conversion, we should enforce uppercase for all future enum types via a linting rule.